add csharp file, project, and open it up in rider.
create a C# library project, and a unit test project attached to it.

There doesn't appear to be any good unit testing solution for godot 4 with C#! apparently they have some of the same support problems.

add a simple model for what could be a single unit in the crowd